Added opencv_contrib_world module and fixed some compilation warnings on iOS

+++ modules/datasets/src/or_sun.cpp
@@ -98,7 +98,7 @@ void OR_sunImp::loadDatasetPart(const string &path, vector< Ptr<Object> > &datas
                 curr->label = (*it).second;
             } else
             {
-                curr->label = pathLabel.size();
+                curr->label = (int)pathLabel.size();
                 pathLabel.insert(make_pair(labelStr, curr->label));
                 paths.push_back(labelStr);
             }
